remove schema file
